Crate aws_smithy_observability

Smithy Observability

Modules

global

Utilities for interacting with the currently set GlobalTelemetryProvider

instruments

Instruments are used to record values for metrics.

meter

Metrics are used to gain insight into the operational performance and health of a system in real time.

Structs

Attributes

Structured telemetry metadata.

GlobalTelemetryProviderError

An simple error to represent issues with the global crate::TelemetryProvider.

ObservabilityError

An error in the SDKs Observability providers

TelemetryProvider

A struct to hold the various types of telemetry providers.

TelemetryProviderBuilder

A builder for TelemetryProvider.

Enums

AttributeValue

The valid types of values accepted by Attributes.

ErrorKind

The types of errors associated with ObservabilityError

Traits

Context

A cross cutting concern for carrying execution-scoped values across API boundaries (both in-process and distributed).

ContextManager

Keeps track of the current Context.

Scope

Delineates a logical scope that has some beginning and end (e.g. a function or block of code).